refactor(pwa): split cloud sync sign-in flow
Summary:
- Extract the cloud sync sign-in form branches into a focused component.
- Move account polling, cache hydration, and device activation state into a
  dedicated hook.
- Refresh Lingui source references after moving translated JSX.

Rationale:
- The route still mixed rendering, account state polling, and device switch
  orchestration after the first decomposition. This keeps the route focused on
  page-level wiring while preserving the intended first-login cloud settings
  creation behavior.

Tests:
- vp run check --fix
- vp run lingui:extract
