+++ packages/pwa/src/routes/settings_.cloud-sync.tsx
@@ -1,6 +1,5 @@
 import { Trans } from "@lingui/react/macro";
 import { t } from "@lingui/core/macro";
-import { isValidDocumentId } from "@automerge/automerge-repo/slim";
 import { createFileRoute, useLocation, useNavigate, useRouter } from "@tanstack/react-router";
 import { useEffect, useRef, useState, type FormEvent } from "react";
 import { AnimatePresence } from "motion/react";
@@ -8,15 +7,15 @@ import { toast } from "sonner";
 import { BackButton } from "#src/components/BackButton.js";
 import { CloudSyncAccountSettings } from "#src/components/CloudSyncAccountSettings.tsx";
 import {
-  AuthButtonIcon,
+  CloudSyncSignInForm,
+  type AuthPendingAction,
+} from "#src/components/CloudSyncSignInForm.tsx";
+import {
   AuthCallbackErrorDialog,
-  AuthErrorAlert,
   CloudActionConfirmationDialog,
-  CloudAuthLoadingState,
   CloudSyncSignInDialog,
   CloudSyncSwitchDialog,
   DeleteAccountDialog,
-  MagicLinkSentState,
   SignInSuccessOverlay,
   type CloudActionDialogType,
 } from "#src/components/CloudSyncDialogs.tsx";
@@ -33,28 +32,15 @@ import {
   requestPasswordResetEmail,
   signInWithSocialAuthAccount,
   type AuthSessionUser,
-  type LinkedAuthAccount,
   type SocialAuthProvider,
 } from "#src/lib/auth-client.ts";
-import {
-  clearCachedCloudUserSettings,
-  fetchCloudUserSettings,
-  getCloudUserSettingsInput,
-  saveCloudUserSettings,
-  writeCachedCloudUserSettings,
-  type CloudUserSettings,
-} from "#src/lib/cloudSyncSettings.ts";
+import { clearCachedCloudUserSettings } from "#src/lib/cloudSyncSettings.ts";
 import {
   clearCachedCloudAccountState,
-  hasLocalPartyListData,
   isEmailFieldErrorMessage,
   isPasswordFieldErrorMessage,
-  readCachedCloudAccountState,
-  writeCachedCloudAccountState,
 } from "#src/lib/cloudSyncRouteState.ts";
-import { Button } from "#src/ui/Button.tsx";
-import { Icon } from "#src/ui/Icon.tsx";
-import { AppTextField } from "#src/ui/fields/TextField.js";
+import { useCloudSyncAccountState } from "#src/hooks/useCloudSyncAccountState.ts";
 import { usePartyList } from "#src/hooks/usePartyList.js";
 import { Settings } from "#src/routes/settings.tsx";
 import { closeRouteState } from "#src/lib/navigationHistory.ts";
@@ -71,23 +57,12 @@ const SIGN_IN_SUCCESS_ANIMATION_MS = 1200;
 const SIGN_IN_SUCCESS_EXIT_ANIMATION_MS = 180;
 const DIALOG_EXIT_ANIMATION_MS = 180;
 const PASSWORD_SIGN_IN_ENABLE_DELAY_MS = 250;
-const CLOUD_ACCOUNT_STATE_POLL_INTERVAL_MS = 30_000;
 const DELETE_ACCOUNT_CONFIRMATION_TEXT = "delete account";
-const AUTH_SECONDARY_BUTTON_CLASS_NAME =
-  "h-9 text-sm font-medium text-accent-700 dark:text-accent-50";
 
 interface CloudSyncSearchParams {
   auth?: "success";
   error?: string;
 }
-
-type AuthPendingAction =
-  | "apple"
-  | "google"
-  | "magic-link"
-  | "password"
-  | "password-reset"
-  | "sign-out";
 
 function CloudSyncSettings() {
   const { partyList } = usePartyList();
@@ -111,28 +86,32 @@ function CloudSyncSettings() {
   const [isSignInSuccessVisible, setIsSignInSuccessVisible] = useState(auth === "success");
   const [magicLinkMessage, setMagicLinkMessage] = useState<string | null>(null);
   const [passwordResetMessage, setPasswordResetMessage] = useState<string | null>(null);
-  const [linkedAccounts, setLinkedAccounts] = useState<LinkedAuthAccount[]>([]);
-  const [cloudSettings, setCloudSettings] = useState<CloudUserSettings | null>(null);
   const [authPendingAction, setAuthPendingAction] = useState<AuthPendingAction | null>(null);
   const [isSignInDialogOpen, setIsSignInDialogOpen] = useState(true);
-  const [isCloudSyncSwitchOpen, setIsCloudSyncSwitchOpen] = useState(false);
   const [cloudActionDialog, setCloudActionDialog] = useState<CloudActionDialogType | null>(null);
   const [authCallbackDialogError, setAuthCallbackDialogError] = useState<string | null>(null);
   const [isDeleteAccountDialogOpen, setIsDeleteAccountDialogOpen] = useState(false);
   const [deleteAccountConfirmation, setDeleteAccountConfirmation] = useState("");
   const [deleteAccountError, setDeleteAccountError] = useState<string | null>(null);
   const [isDeleteAccountPending, setIsDeleteAccountPending] = useState(false);
-  const partyListRef = useRef(partyList);
   const isSignInSuccessVisibleRef = useRef(isSignInSuccessVisible);
-  const linkedProviderIds = new Set(linkedAccounts.map((account) => account.providerId));
   const isAuthPending = authPendingAction !== null;
-  const hasPasswordAccount = linkedProviderIds.has("credential");
   const canDeleteAccount =
     deleteAccountConfirmation.trim().toLowerCase() === DELETE_ACCOUNT_CONFIRMATION_TEXT;
-
-  useEffect(() => {
-    partyListRef.current = partyList;
-  }, [partyList]);
+  const {
+    activateCloudSyncOnDevice,
+    clearCloudSyncState,
+    hasPasswordAccount,
+    isCloudSyncSwitchOpen,
+    linkedProviderIds,
+    saveLinkedAccounts,
+    setIsCloudSyncSwitchOpen,
+  } = useCloudSyncAccountState({
+    isSignInSuccessVisibleRef,
+    onCloudDataActivated,
+    partyList,
+    userId,
+  });
 
   useEffect(() => {
     isSignInSuccessVisibleRef.current = isSignInSuccessVisible;
@@ -152,125 +131,9 @@ function CloudSyncSettings() {
 
   useEffect(() => {
     if (!userId) {
-      setLinkedAccounts([]);
-      setCloudSettings(null);
-      setIsCloudSyncSwitchOpen(false);
       setCloudActionDialog(null);
-      return;
     }
-
-    let isCancelled = false;
-    const currentUserId = userId;
-    const cachedAccountState = readCachedCloudAccountState(currentUserId);
-
-    if (cachedAccountState) {
-      setLinkedAccounts(cachedAccountState.linkedAccounts);
-      setCloudSettings(cachedAccountState.cloudSettings);
-    } else {
-      setLinkedAccounts([]);
-      setCloudSettings(null);
-    }
-
-    void loadAccountState({
-      showErrorToast: !cachedAccountState,
-      showStartToast: true,
-    });
-
-    const intervalId = window.setInterval(() => {
-      void loadAccountState({ showErrorToast: false, showStartToast: false });
-    }, CLOUD_ACCOUNT_STATE_POLL_INTERVAL_MS);
-
-    return () => {
-      isCancelled = true;
-      window.clearInterval(intervalId);
-    };
-
-    async function loadAccountState({
-      showErrorToast,
-      showStartToast,
-    }: {
-      showErrorToast: boolean;
-      showStartToast: boolean;
-    }) {
-      try {
-        const currentPartyList = partyListRef.current;
-        const [accounts, { settings }] = await Promise.all([
-          fetchLinkedAuthAccounts(),
-          fetchCloudUserSettings(),
-        ]);
-
-        if (isCancelled) {
-          return;
-        }
-
-        let activeSettings = settings;
-
-        if (!activeSettings) {
-          const { settings: savedSettings } = await saveCloudUserSettings(
-            getCloudUserSettingsInput(currentPartyList),
-          );
-
-          if (isCancelled) {
-            return;
-          }
-
-          activeSettings = savedSettings;
-          if (showStartToast) {
-            toast.success(t`trizum cloud started`);
-          }
-        }
-
-        setLinkedAccounts(accounts);
-        setCloudSettings(activeSettings);
-        writeCachedCloudUserSettings(currentUserId, activeSettings);
-        writeCachedCloudAccountState(currentUserId, {
-          cloudSettings: activeSettings,
-          linkedAccounts: accounts,
-        });
-
-        if (!activeSettings) {
-          return;
-        }
-
-        if (!isValidDocumentId(activeSettings.partyListDocumentId)) {
-          toast.error(t`trizum cloud data is invalid`);
-          return;
-        }
-
-        if (activeSettings.partyListDocumentId === currentPartyList.id) {
-          return;
-        }
-
-        if (hasLocalPartyListData(currentPartyList)) {
-          setIsCloudSyncSwitchOpen(true);
-          return;
-        }
-
-        localStorage.setItem("partyListId", activeSettings.partyListDocumentId);
-        setCloudSettings(activeSettings);
-        setIsCloudSyncSwitchOpen(false);
-        writeCachedCloudUserSettings(currentUserId, activeSettings);
-        writeCachedCloudAccountState(currentUserId, {
-          cloudSettings: activeSettings,
-          linkedAccounts: accounts,
-        });
-        toast.success(t`trizum cloud enabled on this device`);
-
-        if (isSignInSuccessVisibleRef.current) {
-          window.setTimeout(() => {
-            void navigate({ to: "/", replace: true });
-          }, SIGN_IN_SUCCESS_ANIMATION_MS + SIGN_IN_SUCCESS_EXIT_ANIMATION_MS);
-          return;
-        }
-
-        void navigate({ to: "/", replace: true });
-      } catch {
-        if (!isCancelled && showErrorToast) {
-          toast.error(t`Could not load trizum cloud state`);
-        }
-      }
-    }
-  }, [navigate, userId]);
+  }, [userId]);
 
   useEffect(() => {
     if (auth === "success") {
@@ -492,13 +355,7 @@ function CloudSyncSettings() {
       }
 
       const accounts = await fetchLinkedAuthAccounts();
-      setLinkedAccounts(accounts);
-      if (userId) {
-        writeCachedCloudAccountState(userId, {
-          cloudSettings,
-          linkedAccounts: accounts,
-        });
-      }
+      saveLinkedAccounts(accounts);
       toast.success(t`Sign-in method connected`);
     } catch {
       toast.error(t`Could not connect sign-in method`);
@@ -561,8 +418,7 @@ function CloudSyncSettings() {
     try {
       await authClient.signOut();
       clearNativeAuthToken();
-      setCloudSettings(null);
-      setLinkedAccounts([]);
+      clearCloudSyncState();
       await session.refetch();
       toast.success(t`Signed out`);
       void navigate({ to: "/settings", replace: true });
@@ -571,44 +427,6 @@ function CloudSyncSettings() {
     } finally {
       setAuthPendingAction(null);
     }
-  }
-
-  function activateCloudSyncOnDevice(settings: CloudUserSettings | null = cloudSettings) {
-    if (!settings) {
-      toast.message(t`trizum cloud is not set up yet`);
-      return;
-    }
-
-    if (!isValidDocumentId(settings.partyListDocumentId)) {
-      toast.error(t`trizum cloud data is invalid`);
-      return;
-    }
-
-    if (settings.partyListDocumentId === partyList.id) {
-      toast.message(t`This device is already using trizum cloud`);
-      return;
-    }
-
-    localStorage.setItem("partyListId", settings.partyListDocumentId);
-    setCloudSettings(settings);
-    setIsCloudSyncSwitchOpen(false);
-    if (userId) {
-      writeCachedCloudUserSettings(userId, settings);
-      writeCachedCloudAccountState(userId, {
-        cloudSettings: settings,
-        linkedAccounts,
-      });
-    }
-    toast.success(t`trizum cloud enabled on this device`);
-
-    if (isSignInSuccessVisible) {
-      window.setTimeout(() => {
-        void navigate({ to: "/", replace: true });
-      }, SIGN_IN_SUCCESS_ANIMATION_MS + SIGN_IN_SUCCESS_EXIT_ANIMATION_MS);
-      return;
-    }
-
-    void navigate({ to: "/", replace: true });
   }
 
   function openDeleteAccountDialog() {
@@ -633,9 +451,7 @@ function CloudSyncSettings() {
         clearCachedCloudUserSettings(userId);
         clearCachedCloudAccountState(userId);
       }
-      setCloudSettings(null);
-      setLinkedAccounts([]);
-      setIsCloudSyncSwitchOpen(false);
+      clearCloudSyncState();
       setIsDeleteAccountDialogOpen(false);
       setDeleteAccountConfirmation("");
       await session.refetch();
@@ -661,6 +477,17 @@ function CloudSyncSettings() {
     }, DIALOG_EXIT_ANIMATION_MS);
   }
 
+  function onCloudDataActivated(shouldDelay: boolean) {
+    if (shouldDelay) {
+      window.setTimeout(() => {
+        void navigate({ to: "/", replace: true });
+      }, SIGN_IN_SUCCESS_ANIMATION_MS + SIGN_IN_SUCCESS_EXIT_ANIMATION_MS);
+      return;
+    }
+
+    void navigate({ to: "/", replace: true });
+  }
+
   if (!user) {
     return (
       <div className="relative min-h-full">
@@ -673,201 +500,52 @@ function CloudSyncSettings() {
           onOpenChange={closeSignInDialog}
           showHeader={!magicLinkMessage && auth !== "success" && !isSignInSuccessVisible}
         >
-          {auth === "success" || isSignInSuccessVisible ? (
-            <CloudAuthLoadingState />
-          ) : isPasswordResetMode ? (
-            <form className="flex flex-col gap-4" onSubmit={onPasswordResetSubmit}>
-              {authError ? <AuthErrorAlert message={authError} /> : null}
-              <AppTextField
-                errorMessage={authEmailError ?? undefined}
-                isDisabled={isAuthPending}
-                isInvalid={Boolean(authEmailError)}
-                isRequired
-                label={t`Email`}
-                onChange={onAuthEmailChange}
-                type="email"
-                value={authEmail}
-              />
-              {passwordResetMessage ? (
-                <p className="text-sm text-accent-700 dark:text-accent-50">
-                  {passwordResetMessage}
-                </p>
-              ) : null}
-              <Button color="accent" isDisabled={isAuthPending} type="submit">
-                <span className="flex items-center gap-2">
-                  <AuthButtonIcon
-                    icon="lucide.mail"
-                    isPending={authPendingAction === "password-reset"}
-                  />
-                  <Trans>Send password link</Trans>
-                </span>
-              </Button>
-              <Button
-                color="transparent"
-                isDisabled={isAuthPending}
-                onPress={() => {
-                  setIsPasswordResetMode(false);
-                  clearAuthFeedback();
-                }}
-                type="button"
-              >
-                <Trans>Back to sign in</Trans>
-              </Button>
-            </form>
-          ) : magicLinkMessage ? (
-            <MagicLinkSentState
-              email={authEmail}
-              message={magicLinkMessage}
-              onTryAgain={() => {
-                clearAuthFeedback();
-                setIsPasswordLoginMode(false);
-              }}
-            />
-          ) : (
-            <>
-              <div className="flex flex-col gap-3">
-                <Button
-                  color="input-like"
-                  isDisabled={isAuthPending}
-                  onPress={() => {
-                    void onSocialSignIn("apple");
-                  }}
-                >
-                  <span className="flex items-center gap-2">
-                    <AuthButtonIcon icon="brand.apple" isPending={authPendingAction === "apple"} />
-                    <Trans>Continue with Apple</Trans>
-                  </span>
-                </Button>
-                <Button
-                  color="input-like"
-                  isDisabled={isAuthPending}
-                  onPress={() => {
-                    void onSocialSignIn("google");
-                  }}
-                >
-                  <span className="flex items-center gap-2">
-                    <AuthButtonIcon
-                      icon="brand.google"
-                      isPending={authPendingAction === "google"}
-                    />
-                    <Trans>Continue with Google</Trans>
-                  </span>
-                </Button>
-              </div>
-
-              <div className="flex items-center gap-3 text-xs font-medium uppercase text-accent-600 dark:text-accent-300">
-                <span className="h-px flex-1 bg-accent-200 dark:bg-accent-700" />
-                <Trans>or use email</Trans>
-                <span className="h-px flex-1 bg-accent-200 dark:bg-accent-700" />
-              </div>
-
-              {isPasswordLoginMode ? (
-                <form className="flex flex-col gap-4" onSubmit={onPasswordSignInSubmit}>
-                  {authError ? <AuthErrorAlert message={authError} /> : null}
-                  <AppTextField
-                    errorMessage={authEmailError ?? undefined}
-                    isDisabled={isAuthPending}
-                    isInvalid={Boolean(authEmailError)}
-                    isRequired
-                    label={t`Email`}
-                    onChange={onAuthEmailChange}
-                    type="email"
-                    value={authEmail}
-                  />
-                  <AppTextField
-                    errorMessage={authPasswordError ?? undefined}
-                    isDisabled={isAuthPending}
-                    isInvalid={Boolean(authPasswordError)}
-                    isRequired
-                    label={t`Password`}
-                    minLength={8}
-                    onChange={onAuthPasswordChange}
-                    type="password"
-                    value={authPassword}
-                  />
-                  <Button
-                    className="font-semibold"
-                    color="accent"
-                    isDisabled={isAuthPending || !isPasswordSignInEnabled}
-                    type="submit"
-                  >
-                    <span className="flex items-center gap-2">
-                      <AuthButtonIcon
-                        icon="lucide.log-in"
-                        isPending={authPendingAction === "password"}
-                      />
-                      <Trans>Sign in with password</Trans>
-                    </span>
-                  </Button>
-                  <Button
-                    color="input-like"
-                    isDisabled={isAuthPending}
-                    onPress={() => {
-                      setIsPasswordLoginMode(false);
-                      setIsPasswordSignInEnabled(true);
-                      clearAuthFeedback();
-                    }}
-                    type="button"
-                  >
-                    <span className="flex items-center gap-2">
-                      <Icon icon="lucide.mail" width={18} height={18} />
-                      <Trans>Sign in with magic link</Trans>
-                    </span>
-                  </Button>
-                  <Button
-                    className={AUTH_SECONDARY_BUTTON_CLASS_NAME}
-                    color="transparent"
-                    isDisabled={isAuthPending}
-                    onPress={() => {
-                      setIsPasswordResetMode(true);
-                      clearAuthFeedback();
-                    }}
-                    type="button"
-                  >
-                    <Trans>Forgot password?</Trans>
-                  </Button>
-                </form>
-              ) : (
-                <form className="flex flex-col gap-4" onSubmit={onMagicLinkSubmit}>
-                  {authError ? <AuthErrorAlert message={authError} /> : null}
-                  <AppTextField
-                    errorMessage={authEmailError ?? undefined}
-                    isDisabled={isAuthPending}
-                    isInvalid={Boolean(authEmailError)}
-                    isRequired
-                    label={t`Email`}
-                    onChange={onAuthEmailChange}
-                    type="email"
-                    value={authEmail}
-                  />
-                  <Button color="accent" isDisabled={isAuthPending} type="submit">
-                    <span className="flex items-center gap-2">
-                      <AuthButtonIcon
-                        icon="lucide.mail"
-                        isPending={authPendingAction === "magic-link"}
-                      />
-                      <Trans>Email me a sign-in link</Trans>
-                    </span>
-                  </Button>
-                  <Button
-                    color="input-like"
-                    isDisabled={isAuthPending}
-                    onPress={() => {
-                      setIsPasswordLoginMode(true);
-                      setIsPasswordSignInEnabled(false);
-                      clearAuthFeedback();
-                    }}
-                    type="button"
-                  >
-                    <span className="flex items-center gap-2">
-                      <Icon icon="lucide.key-round" width={18} height={18} />
-                      <Trans>Sign in with password</Trans>
-                    </span>
-                  </Button>
-                </form>
-              )}
-            </>
-          )}
+          <CloudSyncSignInForm
+            auth={auth}
+            authEmail={authEmail}
+            authEmailError={authEmailError}
+            authError={authError}
+            authPassword={authPassword}
+            authPasswordError={authPasswordError}
+            authPendingAction={authPendingAction}
+            isAuthPending={isAuthPending}
+            isPasswordLoginMode={isPasswordLoginMode}
+            isPasswordResetMode={isPasswordResetMode}
+            isPasswordSignInEnabled={isPasswordSignInEnabled}
+            isSignInSuccessVisible={isSignInSuccessVisible}
+            magicLinkMessage={magicLinkMessage}
+            onAuthEmailChange={onAuthEmailChange}
+            onAuthPasswordChange={onAuthPasswordChange}
+            onBackToSignIn={() => {
+              setIsPasswordResetMode(false);
+              clearAuthFeedback();
+            }}
+            onForgotPassword={() => {
+              setIsPasswordResetMode(true);
+              clearAuthFeedback();
+            }}
+            onMagicLinkSubmit={onMagicLinkSubmit}
+            onPasswordResetSubmit={onPasswordResetSubmit}
+            onPasswordSignInSubmit={onPasswordSignInSubmit}
+            onSocialSignIn={(provider) => {
+              void onSocialSignIn(provider);
+            }}
+            onTryAnotherEmail={() => {
+              clearAuthFeedback();
+              setIsPasswordLoginMode(false);
+            }}
+            onUseMagicLink={() => {
+              setIsPasswordLoginMode(false);
+              setIsPasswordSignInEnabled(true);
+              clearAuthFeedback();
+            }}
+            onUsePassword={() => {
+              setIsPasswordLoginMode(true);
+              setIsPasswordSignInEnabled(false);
+              clearAuthFeedback();
+            }}
+            passwordResetMessage={passwordResetMessage}
+          />
         </CloudSyncSignInDialog>
       </div>
     );
